+++ tests/plot/test_plot.py
@@ -151,7 +151,9 @@ def test_plot_2d_graph(setup_basic_pixel_dataset, component, marker, show_b_node
 def test_plot_colocalization_heatmap(setup_basic_pixel_dataset):
     np.random.seed(0)
     pxl_data, *_ = setup_basic_pixel_dataset
-    fig, _ = plot_colocalization_heatmap(pxl_data.colocalization)
+    fig, _ = plot_colocalization_heatmap(
+        pxl_data.colocalization, value_column="pearson"
+    )
     return fig
 
 
@@ -172,12 +174,13 @@ def test_plot_colocalization_diff_heatmap(setup_basic_pixel_dataset):
     colocalization_data.loc[6] = ["CD3", "CD19", 0.7, "PXLCMP0000003"]
     fig, _ = plot_colocalization_diff_heatmap(
         colocalization_data,
-        target="PXLCMP0000003",
+        targets="PXLCMP0000003",
         reference="PXLCMP0000002",
         contrast_column="component",
-        use_z_score=False,
+        value_column="pearson",
+        min_log_p=0,
     )
-    return fig
+    return fig["PXLCMP0000003"]
 
 
 @pytest.mark.mpl_image_compare(
@@ -197,10 +200,35 @@ def test_plot_colocalization_diff_volcano(setup_basic_pixel_dataset):
     colocalization_data.loc[6] = ["CD3", "CD19", 0.7, "PXLCMP0000003"]
     fig, _ = plot_colocalization_diff_volcano(
         colocalization_data,
-        target="PXLCMP0000003",
+        targets="PXLCMP0000003",
         reference="PXLCMP0000002",
         contrast_column="component",
-        use_z_score=False,
+        value_column="pearson",
+        min_log_p=-1,
+    )
+    return fig
+
+
+@pytest.mark.mpl_image_compare(
+    deterministic=True,
+    baseline_dir="../snapshots/test_plot/test_plot_colocalization_diff_volcano_multiple",
+)
+def test_plot_colocalization_diff_volcano_multiple(setup_basic_pixel_dataset):
+    np.random.seed(0)
+    pxl_data, *_ = setup_basic_pixel_dataset
+    colocalization_data = pxl_data.colocalization
+    colocalization_data.loc[5] = [
+        "CD3",
+        "CD19",
+        0.5,
+        "PXLCMP0000002",
+    ]  # Adding a new pair of colocalization data as the volcano needs at least 2 rows
+    colocalization_data.loc[6] = ["CD3", "CD19", 0.7, "PXLCMP0000003"]
+    fig, _ = plot_colocalization_diff_volcano(
+        colocalization_data,
+        reference="PXLCMP0000002",
+        contrast_column="component",
+        value_column="pearson",
         min_log_p=-1,
     )
     return fig